% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{auth_check}
\alias{auth_check}
\title{Verify the connection with the Transifex API}
\usage{
auth_check(req)
}
\arguments{
\item{req}{\code{\link[httr2]{request}}}

\item{verbose}{\code{lgl} whether to include verbose output in console. See \code{\link[httr2]{req_verbose}}}
}
\value{
\code{lgl}
}
\description{
TRANSIFEX_API_SECRET must be set up as an Environment variable for this function to return \code{TRUE}
}
\examples{
\dontrun{
# Assumes TRANSIFEX_API_SECRET env var is set-up with API Key
  if (auth_check())
    print("Transifex API Key is connected")
}
}
