% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.i18n.R
\name{shiny.i18n_json}
\alias{shiny.i18n_json}
\title{Construct a shiny.i18n compatible JSON file}
\usage{
shiny.i18n_json(
  tag_any = NULL,
  tag_all = NULL,
  tag_query = NULL,
  file = NULL
)
}
\arguments{
\item{tag_any}{\code{chr} Tags, combined via ANY}

\item{tag_all}{\code{chr} Tags, combined via ALL}

\item{tag_query}{\code{chr} See \href{Get Resource Strings Collection}{https://developers.transifex.com/reference/get_resource-strings} parameters for documentation}

\item{file}{\code{chr} path of file to write json}
}
\value{
\code{chr} JSON
}
\description{
Construct a shiny.i18n compatible JSON file
}
\examples{
\dontrun {
(json <- shiny.i18n_json(tag_any = "rob_vul_metric"))
 shiny.i18n_json(tag_any = "rob_vul_metric", file = "test.json")
}
}
