% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{req_authorize}
\alias{req_authorize}
\title{Authorize an API Request}
\usage{
req_authorize(req)
}
\arguments{
\item{req}{\code{\link[httr2]{request}}}
}
\value{
\code{\link[httr2]{request}}
}
\description{
TRANSIFEX_API_SECRET must be set up as an Environment variable for this function to function properly
}
